TriePack v1.0.0: full implementation with JSON, docs, and bitstream API improvements
Implement all core libraries (bitstream, trie codec, JSON), 22 passing tests,
6 working examples, comprehensive documentation, and GitHub Pages site.

Key changes in this release:
- Core trie encoder/decoder with typed values and CRC-32 integrity
- JSON library with recursive-descent parser and dot-path flattening
- Bitstream API: uint8_t width params changed to unsigned int for 32-bit efficiency
- Added tp_bs_write_bits_signed() for signed arbitrary-width field writes
- uint64_t portability fallback for platforms without native 64-bit support
- Comprehensive bitstream guide for sub-byte integer I/O
- Weird bit-width tests (1-64 bits, signed and unsigned)
- C++ RAII wrappers for all three library layers
- Jekyll-based documentation site with format specs and API reference

+++ src/bitstream/bitstream_varint.c
@@ -13,7 +13,7 @@ tp_result tp_bs_read_varint_u(tp_bitstream_reader *r, uint64_t *out)
         return TP_ERR_INVALID_PARAM;
 
     uint64_t val = 0;
-    uint8_t shift = 0;
+    unsigned int shift = 0;
 
     for (int group = 0; group < TP_VARINT_MAX_GROUPS; group++) {
         if (r->pos + 8 > r->bit_len)
